+++ src/components/Settings/index.scss
